requirements.txt
loguru==0.7.2
numpy==2.0.2
openai==1.43.0
pandas==2.2.2
pre-commit==3.8.0
python-docx==1.1.2
python-dotenv==1.0.1
scikit-learn==1.5.1
scipy==1.13.1
streamlit==1.38.0
tqdm==4.66.5
watchdog==5.0.2
